refactor: move system configs
